# openapi/openapi3-v1.yaml
...
openapi: 3.0.3
info:
  title: Contract Manager
  description: Contract Manager HTTP APIs for interaction with the Dataspace connector.
  version: '1.1'
  license:
    name: European Union Public License (EUPL) 1.2
    url: https://interoperable-europe.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_en.pdf
  contact:
    name: Simpl Programme
    email: cnect-simpl@ec.europa.eu
    url: https://simpl-programme.ec.europa.eu/
servers:
  - url: https://localhost:8085/contract/v1
    description: Contract Manager
paths:
...
  "/agreements/{contractAgreementId}/definitions/{contractDefinitionId}/status":
    patch:
      tags:
        - contract-agreements
      summary: Confirm signing of the Contract Agreement
      description: Confirm signing of the Contract Agreement by the consumer to the
        provider. Final step in the process.
      operationId: contract#ContractAgreementSignatureConfirmation
      requestBody:
        description: Request body
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/StatusUpdate"
            example:
              status: FINALIZED
        required: true
      parameters:
        - name: contractAgreementId
          in: path
          description: Contract Agreement Id
          required: true
          schema:
            type: string
            format: uuid
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
        - name: contractDefinitionId
          in: path
          description: Contract Definition Id
          required: true
          schema:
            type: string
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
      responses:
        '200':
          description: OK response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ConfirmationResponse"
              example:
                contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
                contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
                status: FINALIZING
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 400
                title: Bad Request
                detail: Invalid body content
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/status"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 401
                title: Unauthorized
                detail: Invalid API Key
        '405':
          description: Method Not allowed
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 405
                title: Method Not Allowed
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd17/definitions/def1/status"
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 500
                title: Server Error
                detail: Server Error
  "/agreements/{contractAgreementId}":
    get:
      tags:
        - contract-agreements
      summary: Get Contract Agreement
      description: Get stored Contract Agreement data from database.
      operationId: contract#ContractAgreement
      parameters:
        - name: contractAgreementId
          in: path
          description: Contract Agreement Id
          required: true
          schema:
            type: string
            format: uuid
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
      responses:
        '200':
          description: OK response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ContractAgreementData"
              example:
                contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
                contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
                consumerSignatureDate: '2024-05-05T10:20:30'
                providerSignatureDate: '2024-05-05T10:20:30'
                status: FINALIZING
                contractNegotiationId: C1-12
                assetId: As5
                providerId: PR1
                consumerId: CNS44
                contractOfferId: C1-12-33
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 400
                title: Bad Request
                detail: Invalid request param
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd16"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 401
                title: Unauthorized
                detail: Invalid API Key
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 404
                title: Not Found
                detail: ContractAgreement not found for contractAgreementId = 7e822f42-f17c-4849-a56f-fbdb5209cd16
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd16"
        '405':
          description: Method Not allowed
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 405
                title: Method Not Allowed
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd16"
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 500
                title: Server Error
                detail: Server Error
  "/agreements/{contractAgreementId}/file":
    get:
      tags:
        - contract-agreements
      summary: Get Contract Agreement File
      description: Get stored Contract Agreement data file from filesystem.
      operationId: contract#ContractAgreementFile
      parameters:
        - name: contractAgreementId
          in: path
          description: Contract Agreement Id
          required: true
          schema:
            type: string
            format: uuid
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
      responses:
        '200':
          description: OK response.
          content:
            text/plain:
              example: "CONTRACT AGREEMENT: 6414065b-9c9c-4d57-9e70-8c8f9a5133cd\nCONTRACT DEFINITION: 7f9d45e2-e9a3-43b7-8403-62ea46b990c6\nASSET: 2\nPROVIDER ID: 3\nPROVIDER SIGNATURE DATE: 2025-09-16 08:06:57\nCONSUMER: 4\nCONSUMER SIGNATURE DATE: 2025-09-16 08:06:57"
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 400
                title: Bad Request
                detail: Invalid request param
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd16/file"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 401
                title: Unauthorized
                detail: Invalid API Key
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 404
                title: Not Found
                detail: ContractAgreement not found for contractAgreementId = 7e822f42-f17c-4849-a56f-fbdb5209cd16
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd16/file"
        '405':
          description: Method Not allowed
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 405
                title: Method Not Allowed
                instance: "/contract/v1/agreements/7e822f42-f17c-4849-a56f-fbdb5209cd16/file"
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 500
                title: Server Error
                detail: Server Error
  "/agreements/{contractNegotiationId}/search":
    post:
      tags:
        - contract-agreements
      summary: Search for existing active Contract Agreement with multiple parameters
      description: Get confirmation if Contract Agreement is stored in database.
      operationId: contract#ContractAgreementSearchParams
      parameters:
        - name: contractNegotiationId
          in: path
          description: Contract Negotiation Id
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9:-]+$"
          example: 0ec61c92-677b-4e97-aef1-ed542e37183b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contractAgreementId:
                  type: string
                  format: uuid
                  description: Contract Agreement Id
                  example: 0ec61c92-677b-4e97-aef1-ed542e37183b
                contractDefinitionId:
                  type: string
                  pattern: "^[A-Za-z0-9:-]+$"
                  description: Contract Definition Id
                  example: 0ec61c92-677b-4e97-aef1-ed542e37183b
                assetId:
                  type: string
                  pattern: "^[A-Za-z0-9:-]+$"
                  description: Contract asset Id
                  example: 123Ab
                providerId:
                  type: string
                  pattern: "^[A-Za-z0-9:-]+$"
                  description: Contract provider Id
                  example: abc12
                consumerId:
                  type: string
                  pattern: "^[A-Za-z0-9:-]+$"
                  description: Contract consumer Id
                  example: XC-12
      responses:
        '200':
          description: OK response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ContractSearchConfirmation"
              example:
                contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
                contractSearchTO:
                  contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
                  contractDefinitionId: 77777777-1179-4cf1-b845-777777777777
                  assetId: A1
                  providerId: P1
                  consumerId: C1
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                type: urn:error:single-custom-exception
                status: 400
                title: Bad Request
                detail: 400 BAD_REQUEST "Validation failure"
                instance: "/contract/v1/agreements/{contractNegotiationId}/search"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 401
                title: Unauthorized
                detail: Invalid API Key
        '405':
          description: Method Not allowed
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 405
                title: Method Not Allowed
                instance: "/contract/v1/agreements/{contractNegotiationId}/search"
        '500':
          description: Server Error
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
              example:
                status: 500
                title: Server Error
                detail: Server Error
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ConfirmationResponse:
      type: object
      required:
        - contractAgreementId
        - contractDefinitionId
        - status
      properties:
        contractAgreementId:
          type: string
          format: uuid
          description: ID of contract agreement
        contractDefinitionId:
          type: string
          description: ID of contract definition
        status:
          type: string
          description: Status of the process, in initiated state.
          example: TERMINATED
          format: string
          enum:
            - INITIATED
            - CREATED
            - FINALIZING
            - FINALIZED
            - TERMINATED
      example:
        contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
        contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
        status: FINALIZING
    ContractSearchConfirmation:
      type: object
      required:
        - contractAgreementId
        - contractDefinitionId
      properties:
        contractAgreementId:
          type: string
          description: ID of contract agreement
        contractDefinitionId:
          type: string
          description: ID of contract definition
      example:
        contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
        contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
    ContractAgreementData:
      type: object
      required:
        - contractAgreementId
        - contractDefinitionId
        - status
      properties:
        contractAgreementId:
          type: string
          format: uuid
          description: ID of contract agreement
        contractDefinitionId:
          type: string
          description: ID of contract definition
        consumerSignatureDate:
          type: string
          description: Consumer signature date
        providerSignatureDate:
          type: string
          description: Provider signature date
        status:
          type: string
          description: Status of the process, in initiated state.
          example: TERMINATED
          format: string
          enum:
            - INITIATED
            - CREATED
            - FINALIZING
            - FINALIZED
            - TERMINATED
        contractNegotiationId:
          type: string
          description: Contract negotiation ID
        assetId:
          type: string
          description: Asset ID
        providerId:
          type: string
          description: Provider ID
        consumerId:
          type: string
          description: Consumer ID
        contractOfferId:
          type: string
          description: Contract offer ID
      example:
        contractAgreementId: 677186f6-1179-4cf1-b845-68d1c00c185f
        contractDefinitionId: 777186f6-1179-4cf1-b845-68d1c00c185f
        consumerSignatureDate: '2024-05-05T10:20:30'
        providerSignatureDate: '2024-05-05T10:20:30'
        status: FINALIZING
        contractNegotiationId: C1-12
        assetId: As5
        providerId: PR1
        consumerId: CNS44
        contractOfferId: C1-12-33
...
    StatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: TERMINATED
          format: string
          enum:
            - FINALIZED
            - TERMINATED
      example:
        status: FINALIZED
    Problem:
      description: A Problem Details object (RFC 9457). Additional properties specific
        to the problem type may be present.
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: about:blank
        title:
          type: string
          description: A short summary of the problem type. Written in English and
            readable for engineers (usually not suited for non technical stakeholders
            and not localized).
          example: Service Unavailable
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server for this
            occurrence of the problem.
          minimum: 400
          maximum: 600
          exclusiveMaximum: true
          example: 503
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of
            the problem
        instance:
          type: string
          format: uri
          description: An absolute URI that identifies the specific occurrence of
            the problem. It may or may not yield further information if dereferenced.
...
security:
  - ApiKeyAuth: []